+++ src/entity/exchange/Exchange.java
@@ -9,7 +9,7 @@ public abstract class Exchange {
 
     protected final Player player1, player2;
     protected long kamas1 = 0, kamas2 = 0;
-    protected ArrayList<World.Couple<Integer, Integer>> items1 = new ArrayList<>(),
+    protected ArrayList<World.Couple<Long, Integer>> items1 = new ArrayList<>(),
             items2 = new ArrayList<>();
     protected boolean ok1, ok2;
 
@@ -18,9 +18,9 @@ public abstract class Exchange {
         this.player2 = player2;
     }
 
-    public static World.Couple<Integer, Integer> getCoupleInList(
-            ArrayList<World.Couple<Integer, Integer>> items, int id) {
-        for (World.Couple<Integer, Integer> couple : items)
+    public static World.Couple<Long, Integer> getCoupleInList(
+            ArrayList<World.Couple<Long, Integer>> items, long id) {
+        for (World.Couple<Long, Integer> couple : items)
             if (couple.first == id)
                 return couple;
         return null;
